requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
torch
transformers
scikit-learn
